Seperate TextUI into Formatter and TextUI so that Formatter can handle the printing and TextUI can handle the inputs.
